requirements.txt
asgiref==3.5.0
cffi==1.15.0
cryptography==36.0.2
Django==4.0.4
django-cors-headers==3.11.0
django-rest-knox==4.2.0
djangorestframework==3.13.1
flake8==4.0.1
mccabe==0.6.1
mypy==0.942
mypy-extensions==0.4.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
